\name{bnr}
\alias{bnr}
\non_function{}
\title{ A data frame of 163 benthic species measured across a transect }
\usage{data(bnr)}
\description{
The \code{bnr} data frame has 103 rows and 163 columns.
Each column is a separate species observed at least once in one of all stations. Several species are very abundant, other are very rare.
}
\source{
 Unpublished dataset.
}
\keyword{datasets}
