\name{extract}
\alias{extract}

\encoding{latin1}

\title{ Extract a subset of the original dataset }
\description{
  `extract' is a generic function for extracting a part of the original dataset according to an analysis...)
}
\usage{
extract(e, n, ...)
}
\arguments{
  \item{e}{ An object from where extraction is performed }
  \item{n}{ The number of elements to extract }
  \item{\dots}{ Additional arguments affecting the extraction }
}
\value{
  A subset of the original dataset contained in the extracted object
}
\seealso{ \code{\link{abund}}, \code{\link{regul}}, \code{\link{tsd}}, \code{\link{turnogram}}, \code{\link{turnpoints}} }

\keyword{ methods }
