\name{.gleissberg.table}
\alias{.gleissberg.table}

\encoding{latin1}

\non_function{}
\title{Table of probabilities according to the Gleissberg distribution}
%\usage{data(gleissberg.table)}
%\arguments{
%  \item{gleissberg.table}{ the dataset to load. Note that this dataset is loaded into the hidden \code{.gleissberg.table} variable (starting with a dot)! }
%}
\description{
The table of probabilities to have 0 to n-2 extrema in a series, for n=3 to 50 (for n > 50, the Gleissberg distribution is approximated with a normal distribution
}
\note{
This dataset is used internally by \code{pgleissberg()}. You do not have to access it directly. See \code{pgleissberg()} for more information }
\seealso{ \code{\link{pgleissberg}} }
\keyword{ datasets }
\keyword{ distribution }
