\name{releve}
\alias{releve}
\non_function{}
\title{ A data frame of six phytoplankton taxa followed in time at one station }
\usage{data(releve)}
\description{
The \code{releve} data frame has 61 rows and 8 columns. Several phytoplankton taxa were numbered in a single station from 1989 to 1992, but at irregular times. 
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Day}{days number, first observation being day 1}
    \item{Date}{strings indicating the date of the observations in "dd/mm/yyyy" format}
    \item{Astegla}{the abundance of Asterionella glacialis}
    \item{Chae}{the abundance of Chaetoceros sp}
    \item{Dity}{the abundance of Ditylum sp}
    \item{Gymn}{the abundance of Gymnodinium sp}
    \item{Melosul}{the abundance of Melosira sulcata + Paralia sulcata}
    \item{Navi}{the abundance of Navicula sp}
  }
}
\source{
 Belin, C. & B. Raffin, 1998. \emph{Les espèces phytoplanctoniques toxiques et nuisibles sur le littoral français de 1984 à 1995, résultats du REPHY (réseau de surveillance du phytoplancton et des phycotoxines)}. Rapport IFREMER, RST.DEL/MP-AO-98-16. IFREMER, France.
}
\keyword{datasets}
